Extract PreferencesAO from SettingsPageAO into top-level class file
